requirements.txt
-r ext/requirements.txt
Flask==1.0.2
Pillow==5.3.0
pytesseract==0.2.5
uwsgi
